onpremise/requirements.txt
opencv-python
azure-storage-blob
torch
supervision
ultralytics
python-dotenv
numpy
